SLEMA-79 Moved files into subdirectories
