+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/add_assessment_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:pg_slema/features/well_being_assessment/presentation/widget/add_assessment_app_bar.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_sleep_form.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_symptoms_form.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_entry.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_well_being_form.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/sleep/assessment_sleep_form.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_entry.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/symptoms/assessment_symptoms_form.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_form.dart';
 import 'package:pg_slema/utils/date/date.dart';
 import 'package:pg_slema/utils/log/logger_mixin.dart';
