+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/forms/sleep/assessment_sleep_quality_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_button.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/sleep_quality.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_button.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/sleep/sleep_quality.dart';
 import 'package:pg_slema/utils/log/logger_mixin.dart';
 
 class AssessmentSleepQualityButton extends StatelessWidget with Logger {
