+++ SlemaForAndroid/lib/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_form.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_form_divider.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_form_title.dart';
-import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/assessment_well_being_selector.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_form_divider.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/common/assessment_form_title.dart';
+import 'package:pg_slema/features/well_being_assessment/presentation/widget/forms/well_being/assessment_well_being_selector.dart';
 import 'package:pg_slema/utils/log/logger_mixin.dart';
 
 class AssessmentWellBeingForm extends StatelessWidget with Logger {
